refactor : contents 개편 (form validation, translater 추가)
