+++ src/components/Mentoring/Create/MentoringCreateForm.tsx
@@ -1,8 +1,9 @@
 import ListBoxInput from '@/components/shared/input/ListBoxInput';
 import TextInput from '@/components/shared/input/TextInput';
 import WrapLabel from '@/components/shared/input/WrapLabel';
-import CareersInput from '@/components/signup-mentor/CareersInput';
-import { fieldCategoryOption, timeScaleOption } from '@/contents';
+import CareersInput from '@/components/signup-informer/CareersInput';
+import { timeScaleOption } from '@/contents/option/scheduleOption';
+import { fieldCategoryOption } from '@/contents/option/mentoringFilteringOption';
 import { ArrowRightIcon } from '@heroicons/react/20/solid';
 import { zodResolver } from '@hookform/resolvers/zod';
 import Link from 'next/link';
@@ -18,35 +19,17 @@ import { useMutation } from 'react-query';
 import CreateSuccessModal from './CreateSuccessModal';
 import useMentorInfo from '@/query/useMentorInfo';
 import Button from '@/components/shared/common/Button';
+import { mentoringCreateFormValidation } from '@/contents/validation/mentoringCreateFormValidation';
 
-const schema = z.object({
-  careers: z.array(
-    z.object({
-      content: z.string(),
-    })
-  ),
-  mentoringName: z.string().min(1, '멘토링 이름은 필수 입력입니다.').max(50, '멘토링 이름은 50자 이하입니다.'),
-  mentoringShortIntro: z.string().min(1, '멘토링 짧은 소개는 필수 입력입니다.').max(200, '짧은 소개는 200자 이하 입니다.'),
-  // mentoringField: z.object({ value: z.string(), name: z.string() }),
-  mentoringField: z.array(z.string()).min(1, '희망 분야는 적어도 한 개 이상 선택해야 합니다.'),
-  mentoringCategory: z.object({ subCategory: z.string(), subValue: z.string() }, { required_error: '카테고리는 필수 입력입니다.' }),
-  mentoringContent: z.string().min(1, '멘토링 소개는 필수 입력입니다.'),
-  price: z.string().min(1, '포인트 가격은 필수 입력입니다.'),
-  timeScale: z.object({ value: z.string(), title: z.string() }),
-  startTimes: z.array(z.string()).min(1, '시간 하나 이상은 필수입력입니다.'),
-});
-
-export type MentoringFormType = z.infer<typeof schema>;
+export type MentoringFormType = z.infer<typeof mentoringCreateFormValidation>;
 
 export default function MentoringCreateForm() {
   const { mutateAsync: mentoringCreateMutate, status: mentoringCreateState } = useMutation(mentoringCreateApi);
   const { data: mentorState, status: mentorStateStatus } = useMentorInfo();
 
-  console.log(mentorState);
-
   const [successModalOpen, setSuccessModalOpen] = useState(false);
   const method = useForm<MentoringFormType>({
-    resolver: zodResolver(schema),
+    resolver: zodResolver(mentoringCreateFormValidation),
     mode: 'onChange',
     defaultValues: {
       mentoringField: [],
