+++ src/components/signup-informer/SignUpSuccessModal.tsx
