+++ src/components/signup/SignUpForm/FirstForm.tsx
@@ -10,30 +10,13 @@ import signUpValidationApi from '@/api/auth/signUpValidationApi';
 import emailSendApi from '@/api/email/emailSendApi';
 import { getErrorMessage } from '@/contents/errorMessage';
 import Button from '@/components/shared/common/Button';
+import { FirstFormValidation } from '@/contents/validation/signUpFormValidation';
 
 interface FirstFormProps {
   nextStep: () => void;
 }
 
-const schema = z.object({
-  email: z.string().email('잘못된 이메일 형식입니다.'),
-  nickName: z.string().min(4, '닉네임은 4글자 이상입니다.').max(10, '닉네임은 10글자 이하입니다.'),
-  aboutPassword: z
-    .object({
-      password: z
-        .string()
-        .min(6, '비밀번호는 8글자 이상입니다.')
-        .regex(
-          /^(?=.*[a-zA-Z])(?=.*\d)(?=.*[$@$!%*?&])[A-Za-z\d$@$!%*?&]{8,10}/,
-          '비밀번호는 영어, 숫자, 특수 문자를 하나 이상 포함해야 합니다.'
-        )
-        .max(20, '비밀번호는 20글자 이하입니다.'),
-      passwordValid: z.string().min(1, '비밀번호 검증은 필수 입력입니다.'),
-    })
-    .refine((data) => data.password === data.passwordValid, { path: ['passwordValid'], message: '비밀번호가 일치하지 않습니다.' }),
-});
-
-export type JoinInfoForm = z.infer<typeof schema>;
+export type JoinInfoForm = z.infer<typeof FirstFormValidation>;
 
 export default function FirstForm({ nextStep }: FirstFormProps) {
   const { mutateAsync: signUpMutate, isLoading: signUpLoading } = useMutation(signUpValidationApi);
@@ -47,7 +30,7 @@ export default function FirstForm({ nextStep }: FirstFormProps) {
     handleSubmit,
     setError,
   } = useForm<JoinInfoForm>({
-    resolver: zodResolver(schema),
+    resolver: zodResolver(FirstFormValidation),
     mode: 'onChange',
     defaultValues: joinInfo,
   });
