+++ src/contents/option/menteeStatusOption.ts
@@ -1,0 +1,26 @@
+export const menteeStatusOption = [
+  {
+    title: '무직',
+    value: '무직',
+  },
+  {
+    title: '현직 주니어',
+    value: '현직 주니어',
+  },
+  {
+    title: '현직 미들',
+    value: '현직 미들',
+  },
+  {
+    title: '경력 휴직',
+    value: '경력 휴직',
+  },
+  {
+    title: '학생',
+    value: '학생',
+  },
+  {
+    title: '기타',
+    value: '기타',
+  },
+];
