+++ src/contents/option/mentoringFilteringOption.ts
@@ -53,61 +53,4 @@ export const fieldCategoryOption = [
   },
 ];
 
-export const timeScaleOption = [
-  {
-    value: '1',
-    title: '00:30',
-  },
-  {
-    value: '2',
-    title: '01:00',
-  },
-  {
-    value: '3',
-    title: '01:30',
-  },
-  {
-    value: '4',
-    title: '02:00',
-  },
-];
-
-export const statusToLabel = {
-  Pending: '승인대기',
-  Assign: '승인완료',
-  MentorRefuse: '승인거부',
-  ExpiredDate: '기한만료',
-  Complete: '완료',
-} as const;
-
 export const sortingOption = ['recent', 'popular', 'starts', 'high_price', 'low_price'];
-export type StatusType = 'Pending' | 'Assign' | 'MentorRefuse' | 'ExpiredDate' | 'Complete';
-
-export const menteeStatusOption = [
-  {
-    title: '무직',
-    value: '무직',
-  },
-  {
-    title: '현직 주니어',
-    value: '현직 주니어',
-  },
-  {
-    title: '현직 미들',
-    value: '현직 미들',
-  },
-  {
-    title: '경력 휴직',
-    value: '경력 휴직',
-  },
-  {
-    title: '학생',
-    value: '학생',
-  },
-  {
-    title: '기타',
-    value: '기타',
-  },
-];
-
-export const yearsToName = (years: number) => (years <= 1 ? '신입' : years <= 3 ? '주니어' : years <= 7 ? '미들' : '시니어');
