+++ src/contents/option/scheduleOption.ts
@@ -1,0 +1,18 @@
+export const timeScaleOption = [
+  {
+    value: '1',
+    title: '00:30',
+  },
+  {
+    value: '2',
+    title: '01:00',
+  },
+  {
+    value: '3',
+    title: '01:30',
+  },
+  {
+    value: '4',
+    title: '02:00',
+  },
+];
