+++ src/contents/translate/getDurationTitle.ts
@@ -1,4 +1,4 @@
-import { timeScaleOption } from '@/contents';
+import { timeScaleOption } from '@/contents/option/scheduleOption';
 
 export default function getDurationTitle(value: string) {
   return timeScaleOption.filter((elem) => elem.value === value)[0].title;
