+++ src/contents/translate/yearToRank.ts
