+++ src/pages/mentoring/[page]/apply.tsx
@@ -1,4 +1,4 @@
-import ManteeInfoForm from '@/components/Mentoring/Apply/ManteeInfoForm';
+import MenteeInfoForm from '@/components/Mentoring/Apply/MenteeInfoForm';
 import MentoringInfoForm from '@/components/Mentoring/Apply/MentoringInfoForm';
 import { NextPage } from 'next';
 import { useState } from 'react';
@@ -11,7 +11,7 @@ const ApplyPage: NextPage = () => {
   return (
     <div className="m-auto max-w-xl px-4">
       {progressStatus === 'first' ? (
-        <ManteeInfoForm onNext={onNext}></ManteeInfoForm>
+        <MenteeInfoForm onNext={onNext}></MenteeInfoForm>
       ) : (
         <MentoringInfoForm onPrev={onPrev}></MentoringInfoForm>
       )}
