+++ src/recoil/form/mentoringApply/menteeFormAtom.tsx
@@ -1,4 +1,4 @@
-import { MentoringApplyMenteeType } from '../../../components/Mentoring/Apply/ManteeInfoForm';
+import { MentoringApplyMenteeType } from '../../../components/Mentoring/Apply/MenteeInfoForm';
 import { atom } from 'recoil';
 
 const menteeFormAtom = atom<MentoringApplyMenteeType>({
